Creating filter_odd_squares and build_pixel_map functions
The filter_odd_squares function removes all the tuples from the grid that have an odd values (the remaining are the tuples that will be "colored" in our identicon). The build_pixel_map function takes this new grid and generates a pixel_map that contains tuples denoting the coordinates of each cell of the grid to be colored.